Rename Playbook "topics" to "chapters"
In the Playbook we have two types of contents: first level contents and
second level contents. Currently the first level contents are called
"sections" throughout all the playbook, but the second level contents
are referred as "topics" or "chapters". In order to level out namings,
we decided to name the second level contents "chapters".